release_update_Q_02

+++ src/custom_allocator.cpp
@@ -1,6 +1,6 @@
 #include "custom_allocator.hpp"
 
-void vs::PrintContainerMap(std::map<int, int, std::less<int>, vs::custom_allocator<std::pair<const int, int>, 10>>& dict) {
+void vs::PrintContainerMap(std::map<size_t, size_t, std::less<size_t>, vs::custom_allocator<std::pair<const size_t, size_t>, 10>>& dict) {
 	for (const auto& node : dict) {
 		std::cout << node.first << " " << node.second << std::endl;
 	}
